Stream responses
This commit ensures the server does not retain the head of the result
sequence. Previously, the behavior was to stream results to the client
but retain the full resultset in memory until traversed. This was due to
a suspected clojure compiler bug demonstrated here:
https://github.com/wkalt/streaming-demo

To get around it, this commit moves the try/catch to a new scope, which
seems to fix the seq retention.